First upload of the insides of the Sublime Text 3 package
This is the first commit of the innards of the Hacknet Snippets package
for Sublime Text 3. At this current moment all the files still need to
be re-checked (yes i checked them once but it's good to recheck) to make
sure that the Tab Triggers match the ones in the README.md and that
there are no spelling errors in any of the tags. A single one and this
package is useless seeing as the XML files won't load up correctly.

Anyway, at this current moment as I re-check them I'm add in the autotab
thing I saw on the docs so that when people trigger the autocomplete,
they can tab through and add in the data they need to.

So it's half that and half the old package.

Speaking of old package I'm popping that in the releases section of this
repo in case anyone feels brave.

Also remember to link to the OG Hacknet AHK Tools repo seeing as that's
how I got my idea.