Removed old package and updated snippets.
Two things:

First I spent the past day updating all the snippets so they use the
autotab function. I gave it a little test run and I feels a lot better
generating the tags and then tabbing through the relevant areas to put
in info. So I'm glad it's working as intented.

Second, I removed the .sublime-package. Mostly because it doesn't need
to be there.

I don't see myself making anymore changes to this unless I fucked up the
spelling somewhere or want to change what the Tab Trigger is for some of
the tags.